parser_service/requirements.txt
celery
redis
# PyPDF2
PyMuPDF
python-pptx
sqlalchemy
psycopg2-binary